server/requirements.txt
flask>=3.0
flask-cors>=4.0
google-genai>=0.8
python-dotenv>=1.0
Pillow>=10.0